First pass as extracting RubyFile
